+++ scripts/devcontainer_postCreateCommand.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+# Devcontainer postCreateCommand.
+# Install dependencies for running this project in GitHub Codespaces.
+
+set -eux
+
+# For git version tagging:
+go install github.com/maykonlsf/semver-cli/cmd/semver@latest